requirements.txt
requests
python-telegram-bot
psycopg2~=2.9.3

pandas~=1.0.5